requirements.txt
# Telegram Bot
python-telegram-bot>=21.0

python-telegram-bot[job-queue]>=21.0

# Anime scraping
animeworld>=1.6.0

# Environment variables
python-dotenv>=1.0.0

# HTTP clients
requests>=2.31.0
httpx>=0.27.0

# HTML parsing (for StreamingCommunity)
beautifulsoup4>=4.12.0

# User agent generation
fake-useragent>=1.4.0

# Date parsing
python-dateutil>=2.8.0

# Console colors
colorama>=0.4.6

# Optional: Streaming Unity API (for gats.py)
# scuapi>=1.0.0

# FastAPI (PWA Backend)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=1.0.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
